New alignmentBins field in BinnedSpikeSet
This field contains information about which bin the startMs and endMs got
aligned to in the final output. It gets appropriately updated with indexing.
The point of having this field is to additionally be able to only grab segments
of spike data that exist (as opposed to going outside the valid bounds), but
retain information about what was meant to be aligned where. That way, if some
trials end earlier or start later, then we're still able to combine them with
other trials. Unfortunate side effect--BinnedSpikeData is more likely to end up
in a list of trials rather than an array because even one trial having a
different number of valid time points will prevent it from becoming an array.
Could consider having a force input that aligns everything to the smallest trial
available... after all, the alignmentBins field will still be there to tell us
how to read things!