New functions for bad channel removal
In both cases, the time points of channels whose firing rate is >3*std of the
whole channel are marked. For one new utility, I check if any of the remaining
(unmarked) timepoints have non-zero (or non-zero after readding the baseline
subtracted) firing rates. If none do, I remove the channel. This helps because
V4 does this sometimes. In another, I check this for each trial individually.
If >90% of the trials have a non-zero firing rate timepoint, keep the channel,
otherwise dump it. It may be noted that the first check is a more stringent
version of the second check (I believe). Until I'm fully satisfied that's the
case, I'm keeping both--but I'll probably transition just to the second, which
actually checks for any non-outlier firing at all in >90% of trials.