List plotting methods adjusted for inputting lists of subsamples
If sessions are subsamples, they all come grouped in a list entry, and now these functions recognize that and keep plotting correctly. This is instead of coming as distinct list entries on the top level list. In addition, the plots now look nicer with more descriptive info and more spacing of the axes and labels... at least matplotlib promises that's what should be happening...