Correct baseline subtraction for object dtype to not change input
Python's pass-by-reference on occasion becomes a killer. In this particular
case, I wanted to output a *copy* of the input array, so I was calling the copy
method of ndarray. However, when the BinnedSpikeSet is of dtype object, the
copy method correctly replaces the references to the top level array, but not
the references of the embedded objects (i.e. the actual binned spike
counts/firing rates)! So when I was baseline subtracting I was actually
affecting the parent binned spike set as well >.> In any case, I actually think
it didn't affect any of my analyses... except for the one where it errored in a
way that it wouldn't have affected previous analyses without erroring. But the
fix was using deepcopy instead of copy to make sure the references of the
embedded objects are also changed, and now it works!